+++ src/libsodium/include/sodium/crypto_box.h
@@ -58,22 +58,22 @@ int crypto_box_beforenm(unsigned char *k, const unsigned char *pk,
 
 SODIUM_EXPORT
 int crypto_box_afternm(unsigned char *c, const unsigned char *m,
-                       unsigned long long mlen, const unsigned char *n,
+                       uint64_t mlen, const unsigned char *n,
                        const unsigned char *k);
 
 SODIUM_EXPORT
 int crypto_box_open_afternm(unsigned char *m, const unsigned char *c,
-                            unsigned long long clen, const unsigned char *n,
+                            uint64_t clen, const unsigned char *n,
                             const unsigned char *k);
 
 SODIUM_EXPORT
 int crypto_box(unsigned char *c, const unsigned char *m,
-               unsigned long long mlen, const unsigned char *n,
+               uint64_t mlen, const unsigned char *n,
                const unsigned char *pk, const unsigned char *sk);
 
 SODIUM_EXPORT
 int crypto_box_open(unsigned char *m, const unsigned char *c,
-                    unsigned long long clen, const unsigned char *n,
+                    uint64_t clen, const unsigned char *n,
                     const unsigned char *pk, const unsigned char *sk);
 
 #ifdef __cplusplus
