+++ src/libsodium/include/sodium/crypto_shorthash_siphash24.h
@@ -18,7 +18,7 @@ SODIUM_EXPORT
 const char * crypto_shorthash_siphash24_primitive(void);
 
 SODIUM_EXPORT
-int crypto_shorthash_siphash24(unsigned char *,const unsigned char *,unsigned long long,const unsigned char *);
+int crypto_shorthash_siphash24(unsigned char *,const unsigned char *,uint64_t,const unsigned char *);
 
 #ifdef __cplusplus
 }
